Add connecting to IMAP
This is a very optimistic implementation to get started. More safeguards
will be added later.

Because the hostname of the IMAP server shouldn't be in source control,
user secrets have been added as well. Where needed, a readme is added to
detail which ones are needed, and how they can be added.

Because this may not be obvious when just the repository is checked out,
also a readme at the root has been added to reference to other ones.