Add retrieving IMAP folder
The goal is hiding the actual implementation, so that the public
interface won't have to change if ever changing away from MailKit. That
is why the constructor is `internal` and a "convenience" `static Create`
method was added.

Because the implementation is by design dependent on each other, it is
OK to not use `MailboxConnection`, but `ImapClient` within
`MailboxFolder`. The reason being is that otherwise any functionality
would have to be replicated in `MailboxConnection`, for no real good
reason. It also makes the code much more dependent on MailKit from the
outside, which is undesirable (unless much more `internal` methods are
created...)