Add fetching of all meta data from a folder
Fetching requires the folder to have been opened, or in other words
"selected". There are different modes, but for now read-write is
mandatory due to always expunging on dispose.

Noticed when testing the exception, that the folder name was incorrect
in the case of a subfolder. This is now fixed on `MailboxConnection`.

Didn't implement yet a way to load the expected set of e-mails into
a specific folder automatically (or when explicitly targeted). For now,
added a sort of placeholder in `README.md` with the remark that 3
e-mails must be in `UnitTestMails` folder.